Move logic into existing example
