+++ edge-functions/ab-testing-statsig/.env.example
@@ -1,2 +1,5 @@
+EDGE_CONFIG=
+EDGE_CONFIG_ITEM_KEY=
 STATSIG_SERVER_API_KEY=
+STATSIG_CONSOLE_API_KEY=
 NEXT_PUBLIC_STATSIG_CLIENT_KEY=
