+++ edge-functions/ab-testing-statsig/README.md
@@ -1,23 +1,23 @@
 ---
-name: A/B Testing with Statsig
+name: Experimentation with Statsig
 slug: ab-testing-statsig
-description: By A/B testing at the edge with Statsig, you'll reduce CLS and improve performance from client-loaded experiments while analyzing impact.
+description: By running experiments at the edge with Statsig, you'll reduce CLS and improve performance from client-loaded experiments while analyzing impact.
 framework: Next.js
 useCase:
   - Edge Functions
   - Documentation
 css: Tailwind
-deployUrl: https://vercel.com/new/clone?repository-url=https://github.com/vercel/examples/tree/main/edge-functions/ab-testing-statsig&project-name=ab-testing-statsig&repository-name=ab-testing-statsig
-demoUrl: https://edge-functions-ab-testing-statsig.vercel.app
+deployUrl: https://vercel.com/new/clone?repository-url=https://github.com/vercel/examples/tree/main/edge-functions/ab-testing-statsig&project-name=ab-testing-statsig&repository-name=ab-testing-statsig&env=STATSIG_SERVER_API_KEY,NEXT_PUBLIC_STATSIG_CLIENT_KEY,STATSIG_CONSOLE_API_KEY,EDGE_CONFIG,EDGE_CONFIG_ITEM_KEY&envDescription=Statsig%20API%20keys%20and%20Edge%20Config%20settings&envLink=https%3A%2F%2Fdocs.statsig.com%2Fguides%2Ffirst-feature
+demoUrl: https://edge-functions-experimentation-statsig.vercel.app
 ---
 
-# A/B testing with Statsig
+# Experimentation with Statsig
 
-This example shows how to do A/B testing using statsig.
+This example shows how to do experimentation using Statsig, leveraging Edge Config.
 
 ## Demo
 
-https://edge-ab-testing-statsig.vercel.app
+https://edge-functions-experimentation-statsig.vercel.app
 
 ## How to Use
 
@@ -27,7 +27,7 @@ You can choose from one of the following two methods to use this repository:
 
 Deploy the example using [Vercel](https://vercel.com?utm_source=github&utm_medium=readme&utm_campaign=vercel-examples):
 
-[![Deploy with Vercel](https://vercel.com/button)](https://vercel.com/new/git/external?repository-url=https://github.com/vercel/examples/tree/main/edge-functions/ab-testing-statsig&project-name=ab-testing-statsig&repository-name=ab-testing-statsig&env=STATSIG_SERVER_API_KEY,NEXT_PUBLIC_STATSIG_CLIENT_KEY&envDescription=API%20keys%20used%20by%20statsig%20on%20the%20server%20and%20the%20client&envLink=https%3A%2F%2Fdocs.statsig.com%2F%23account-sign-up-and-api-key)
+[![Deploy with Vercel](https://vercel.com/button)](https://vercel.com/new/git/external?repository-url=https://vercel.com/new/clone?repository-url=https://github.com/vercel/examples/tree/main/edge-functions/ab-testing-statsig&project-name=ab-testing-statsig&repository-name=ab-testing-statsig&env=STATSIG_SERVER_API_KEY,NEXT_PUBLIC_STATSIG_CLIENT_KEY,STATSIG_CONSOLE_API_KEY,EDGE_CONFIG,EDGE_CONFIG_ITEM_KEY&envDescription=Statsig%20API%20keys%20and%20Edge%20Config%20settings&envLink=https%3A%2F%2Fdocs.statsig.com%2Fguides%2Ffirst-feature)
 
 ### Clone and Deploy
 
@@ -42,12 +42,14 @@ yarn create next-app --example https://github.com/vercel/examples/tree/main/edge
 #### Set up environment variables
 
 Log in to the [Statsig console](https://console.statsig.com/) and navigate to **Settings -> API KEYS**.
-
-Then, copy [.env.example](./env.example) to `.env.local` and fill it in with your Client API keys:
+Then, copy [.env.example](./env.example) to `.env.local` and fill it in with your Client, Server and Console API keys:
 
 ```bash
 cp .env.example .env.local
 ```
+
+Install the [Statsig Vercel Integration](https://vercel.com/integrations/statsig) for your project.
+Then fill in `.env.local` with the provided Edge Config Connection String and Edge Config Item Key.
 
 Next, run Next.js in development mode:
