+++ edge-functions/ab-testing-statsig/lib/constants.ts
@@ -1,5 +1,5 @@
 export const UID_COOKIE = 'uid'
-// This is the fallback experiment in case Statsig doesn't return one
-export const DEFAULT_GROUP = 'a'
 // This is the experiment that will be used to determine the bucket
-export const FLAG = 'statsig_example'
+export const EXPERIMENT = 'statsig_example'
+// Default Experiment Group Fallback
+export const GROUP_PARAM_FALLBACK = 'error_default_bucket'
