+++ edge-functions/ab-testing-statsig/package.json
@@ -14,6 +14,8 @@
     "next": "canary",
     "react": "latest",
     "react-dom": "latest",
+    "statsig-node": "^5.1.0",
+    "statsig-node-vercel": "^0.0.3",
     "statsig-react": "^1.12.0"
   },
   "devDependencies": {
