--- edge-functions/edge-config-statsig/.env.example
@@ -1,5 +1,0 @@
-EDGE_CONFIG=
-EDGE_CONFIG_ITEM_KEY=
-STATSIG_SERVER_API_KEY=
-STATSIG_CONSOLE_API_KEY=
-NEXT_PUBLIC_STATSIG_CLIENT_KEY=
