--- edge-functions/edge-config-statsig/.eslintrc.json
@@ -1,4 +1,0 @@
-{
-  "root": true,
-  "extends": "next/core-web-vitals"
-}
