--- edge-functions/edge-config-statsig/.npmrc
@@ -1,2 +1,0 @@
-# Enabled to avoid deps failing to use next@canary
-legacy-peer-deps=true
