--- edge-functions/edge-config-statsig/lib/constants.ts
@@ -1,5 +1,0 @@
-export const UID_COOKIE = 'uid'
-// This is the experiment that will be used to determine the bucket
-export const EXPERIMENT = 'statsig_example'
-// Default Experiment Group Fallback
-export const GROUP_PARAM_FALLBACK = 'error_default_bucket'
