--- edge-functions/edge-config-statsig/package.json
@@ -1,32 +1,0 @@
-{
-  "repository": "https://github.com/vercel/examples.git",
-  "license": "MIT",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@vercel/examples-ui": "^1.0.4",
-    "js-cookie": "^3.0.1",
-    "next": "canary",
-    "react": "latest",
-    "react-dom": "latest",
-    "statsig-node": "^5.1.0",
-    "statsig-node-vercel": "^0.0.3-beta.1",
-    "statsig-react": "^1.12.0"
-  },
-  "devDependencies": {
-    "@types/js-cookie": "^3.0.1",
-    "@types/node": "^17.0.14",
-    "@types/react": "latest",
-    "autoprefixer": "^10.4.2",
-    "eslint": "^8.16.0",
-    "eslint-config-next": "canary",
-    "postcss": "^8.4.6",
-    "tailwindcss": "^3.0.18",
-    "typescript": "^4.7.2"
-  }
-}
